requirements.txt
bcrypt==4.0.1
colorama==0.4.6
pillow==10.0.0